requirements.txt
Django>=5.2.8
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.7.0
face-recognition>=1.3.0
django-filter>=23.2
reportlab>=4.0.7
openpyxl>=3.1.2
hypothesis>=6.82.0
hypothesis[django]>=6.82.0
gunicorn>=21.2.0
dj-database-url>=2.1.0
whitenoise>=6.5.0